+++ test/endpoints.js
@@ -2,5 +2,7 @@ module.exports = {
     binary:'binary://localhost:6370',
     binarys:'binarys://localhost:6371',
     jsonrpc:'jsonrpc://localhost:6372',
-    jsonrpcs:'jsonrpcs://localhost:6373'
+    jsonrpcs:'jsonrpcs://localhost:6373',
+    dinary:'dinary://localhost:6374',
+    dinarys:'dinarys://localhost:6375'
 };